shuffle queues on fetch to prevent starvation
faktory checks given queues in order, such that consistent and
repetitive fetches to default, critical, low will *always* return a
default job if the default queue is full--regardless of how many jobs
are in the critical queue. Randomizing the order of the requested queues
prevents queue starvation and allows all queues to be worked on. this
does not take priority into account.