move shuffle to file, simplify this.current waits
there was a leak in which awaiting the current job in progress by a
processor would only wait the jobfn itself, not the async process of
ack/failing the job, so during graceful shutdown a race condition
existed in which the job in progress would be awaited, but the
connection pool is shutdown before the processor acks or fails the job
after running the jobfn. This sets the entire dispatch and execution
process into the current awaitable-in-progress so that when awaiting
during graceful shutdown, the entire job execution process, including
eventual ack or fail can complete before draining the connection pool
and disallowing the processor to fully complete the work.