add tests and error wrapping for #2
the processor expecting that if your job returns a rejected promise, that there's an error that it is rejecting with—rather than undefined or a string. In your example, reject('Test') rejects the promise with a string, not an error. When the processor encounters this, it attempts to FAIL your job in the faktory server. FAIL requires an error message and stack trace when sent to Faktory Server, which is why this lib is attempting to call e.message, e.stack.split(...)` and so on.

https://www.pixelstech.net/article/1324645232-A-String-is-not-an-Error

closes #2