gracefully reconnect while working
upgrades faktory-client to 0.5.0 which includes automatic reconnects on
unexpected socket closures. This adds the option for the connection pool
to test that the connection is connected before checking out of the
pool. Timeout for waiting for pool acquisition is 10s. Worker loop will
continue on reconnect without exiting or silently dying.

some tests were added around signal handling.