don't emit('error') on connection
since there are no listeners to 'error' event on the connection and the
connection (pooled) is not exposed to the user, no error handler can be
added to it. This is problematic because 'error' is a special event in
node that, if uncaught, will crash the process.

this changes the connection class to not reemit socket errors on itself,
which would crash the process if uncaught (always) and instead log them
to stderr.

The timeout for connecting has also been increased from 5 to 10 seconds.

fixes #31
fixes #30