remove conn error handler after close
while inducing some failure scenarios, it appeared that there is a race
between destroying a connection due to a connect failure and removing
the error event listener when destroying. The result: an error event is
emitted on connection, but there is no listener for it. This is an
unhandled exception.

related to #70