add default error event handler if none
If no error handlers are added, the error event is an unhandled
exception. An unhandled error may terminate the process in later
eversion of node, so it would be safest to allow an emitted error (like
connection failure during fetch) to be logged instead of disrupting the
process.

related to #70