reduce complexity by removing parser abstraction
removing the layer between connection and redis-parser simplifies the
code a bit and involves one fewer event listener. the promise interface
for connection now looks a bit easier.

the goal here is to reduce moving parts and isolate the connectivity
issues seen when the server disconnects. I've only reproduced the issue
one time, so it's difficult to track down.